refactor(layout): remove the 3D scene
VMX-154 settled that panes stop being off-screen composited surfaces, and
this is the half that has to land first: a perspective free camera can tilt
a pane and a native NSView cannot, so OSR stays load-bearing until the tilt
is gone.

The replacement was already written. scene_user.rs held the orthographic
scene behind cfg(not(feature = "player-mode")), so this deletes the feature
and its true-branch rather than building anything new. It ships today --
vmux_desktop's default is "full", and full listed player-mode -- so the
perspective camera, free camera, bloom and mode transitions were compiled
into every desktop build.

Out: scene.rs, focus_ring.rs, focus_ring.wgsl, window.wgsl, and 46 cfg
sites. scene_user.rs takes the scene module path.

bevy_animation, bevy_animation_macros, bevy_camera_controller,
bevy_post_process and petgraph leave the lockfile, and bevy_pbr no longer
resolves for vmux_desktop at all. bevy_cef keeps its pbr feature for its own
examples, which are now gated behind required-features so the workspace
still builds without it.

InteractionMode survives as a single-variant enum, so behaviour matches the
existing non-player configuration exactly. Collapsing it and removing the
Player menu entry is a follow-up, kept separate because it has no behaviour
change at all.
