+++ crates/page/vmux_layout/Cargo.toml
@@ -9,14 +9,6 @@ build = "../../build_cfg_git.rs"
 [features]
 default = []
 debug = []
-player-mode = [
-    "bevy/bevy_animation",
-    "bevy/bevy_camera_controller",
-    "bevy/bevy_pbr",
-    "bevy/bevy_post_process",
-    "bevy/free_camera",
-    "bevy_cef/pbr",
-]
 
 [lib]
 path = "src/lib.rs"
